Document the demo
