+++ demo/nodejs-domino/another.ts
@@ -1,2 +1,3 @@
+// This file shows that the DOM symbols are available in the global scope
 console.log('typeof window', typeof window)
 console.log('typeof document', typeof document)
