+++ demo/nodejs-domino/custom.ts
@@ -6,46 +6,49 @@ import {
   TrustedTypesEnforcer,
 } from 'trusted-types'
 
-function acceptOnlyTrustedHTML(html: TrustedHTML) {
-  if (!trustedTypes.isHTML(html)) console.log('untrusted html', html)
-  else console.log('trusted html', html)
-}
-
-const policy = trustedTypes.createPolicy('app', { createHTML: (s) => s })
-acceptOnlyTrustedHTML('str' as any)
-acceptOnlyTrustedHTML(policy.createHTML('safe html'))
-
+// create custom DOM API implementation
 const win = domino.createWindow()
 const doc = win.document
 
 const html = doc.createElement('html')
 html.appendChild(doc.createElement('div'))
 html.appendChild(doc.createElement('div'))
-html.appendChild(doc.createElement('div'))
-
 console.log(html.innerHTML)
 
-// no fail
+// no enforcement yet, assignment should work
 html.children[0].innerHTML = 'string'
 console.log(html.innerHTML)
 
-const config = new TrustedTypeConfig(false, true, ['foo'], false)
+// start enforcing mode
+const config = new TrustedTypeConfig(false, true, ['foo', 'default'], false)
 const enforcer = new TrustedTypesEnforcer(config, win)
 enforcer.install()
 const fooPolicy = trustedTypes.createPolicy('foo', { createHTML: (s) => s })
+
+// we expect string assignment to sink to fail
+let caught = false
 try {
   html.children[0].innerHTML = 'string' // should fail
-  throw new Error('Unexpected code path')
 } catch (err) {
-  console.log('Caught unsafe write to innerHTML property')
+  caught = true
 }
+if (caught) console.log('Caught unsafe write to innerHTML property')
+else throw new Error("Didn't catch unsafe write to innerHTML property!")
 
-html.children[0].innerHTML = fooPolicy.createHTML('safeHTML') as any // should work
+// trusted value assignment should pass
+html.children[0].innerHTML = fooPolicy.createHTML('safeHTML') as any
 console.log(html.innerHTML)
 
+// Default policy
+trustedTypes.createPolicy('default', { createHTML: (s) => s + '-default' })
+
+// after we uninstall enforcer, raw assignments should work again
 enforcer.uninstall()
-html.children[0].innerHTML = 'string after uninstall' // should work
+html.children[0].innerHTML = 'string after uninstall'
 console.log(html.innerHTML)
+
+html.children[0].innerHTML = 'string' // should be rewritten to 'string-default'
+console.log(html.children[0].innerHTML)
 
 // it should work even with incomplete enforcer
 const incompleteEnforcer = new TrustedTypesEnforcer(config, {} as Window)
