+++ demo/nodejs-domino/main.ts
@@ -1,4 +1,9 @@
 import domino from 'domino'
+import {
+  trustedTypes,
+  TrustedTypeConfig,
+  TrustedTypesEnforcer,
+} from 'trusted-types'
 
 // expose DOM API on the global scope
 const g = global as any
@@ -8,47 +13,42 @@ Object.entries((domino as any).impl).forEach(([key, value]) => {
   g[key] = value
 })
 
+// console logs whether the global DOM symbols
 import './another'
-import {
-  trustedTypes,
-  TrustedTypeConfig,
-  TrustedTypesEnforcer,
-} from 'trusted-types'
-
-function acceptOnlyTrustedHTML(html: TrustedHTML) {
-  if (!trustedTypes.isHTML(html)) console.log('untrusted html', html)
-  else console.log('trusted html', html)
-}
-
-const policy = trustedTypes.createPolicy('app', { createHTML: (s) => s })
-acceptOnlyTrustedHTML('str' as any)
-acceptOnlyTrustedHTML(policy.createHTML('safe html'))
 
 const html = document.createElement('html')
 html.appendChild(document.createElement('div'))
 html.appendChild(document.createElement('div'))
 html.appendChild(document.createElement('div'))
-
 console.log(html.innerHTML)
 
-// no fail
+// no enforcement yet, assignment should work
 html.children[0].innerHTML = 'string'
 console.log(html.innerHTML)
 
-const config = new TrustedTypeConfig(false, true, ['foo'], false)
+// start enforcing mode
+const config = new TrustedTypeConfig(false, true, ['foo', 'default'], false)
 const enforcer = new TrustedTypesEnforcer(config)
 enforcer.install()
 const fooPolicy = trustedTypes.createPolicy('foo', { createHTML: (s) => s })
+// we expect string assignment to sink to fail
+let caught = false
 try {
   html.children[0].innerHTML = 'string' // should fail
-  throw new Error('Unexpected code path')
 } catch (err) {
-  console.log('Caught unsafe write to innerHTML property')
+  caught = true
 }
+if (caught) console.log('Caught unsafe write to innerHTML property')
+else throw new Error("Didn't catch unsafe write to innerHTML property!")
 
-html.children[0].innerHTML = fooPolicy.createHTML('safeHTML') as any // should work
+// trusted value assignment should pass
+html.children[0].innerHTML = fooPolicy.createHTML('safeHTML') as any
 console.log(html.innerHTML)
 
+// Default policy
+trustedTypes.createPolicy('default', { createHTML: (s) => s + '-default' })
+
+// after we uninstall enforcer, raw assignments should work again
 enforcer.uninstall()
-html.children[0].innerHTML = 'string after uninstall' // should work
+html.children[0].innerHTML = 'string after uninstall'
 console.log(html.innerHTML)
